factor out range_to_text
